+++ simple_history/models.py
@@ -26,6 +26,7 @@ else:  # south configuration for CustomForeignKeyField
     add_introspection_rules(
         [], ["^simple_history.models.CustomForeignKeyField"])
 
+from . import exceptions
 from .manager import HistoryDescriptor
 
 registered_models = {}
@@ -35,10 +36,11 @@ class HistoricalRecords(object):
     thread = threading.local()
 
     def __init__(self, verbose_name=None, bases=(models.Model,),
-                 user_related_name='+', table_name=None):
+                 user_related_name='+', table_name=None, inherit=False):
         self.user_set_verbose_name = verbose_name
         self.user_related_name = user_related_name
         self.table_name = table_name
+        self.inherit = inherit
         try:
             if isinstance(bases, six.string_types):
                 raise TypeError
@@ -49,7 +51,8 @@ class HistoricalRecords(object):
     def contribute_to_class(self, cls, name):
         self.manager_name = name
         self.module = cls.__module__
-        models.signals.class_prepared.connect(self.finalize, sender=cls)
+        self.cls = cls
+        models.signals.class_prepared.connect(self.finalize, weak=False)
         self.add_extra_methods(cls)
 
     def add_extra_methods(self, cls):
@@ -69,6 +72,19 @@ class HistoricalRecords(object):
                 save_without_historical_record)
 
     def finalize(self, sender, **kwargs):
+        try:
+            hint_class = self.cls
+        except AttributeError:  # called via `register`
+            pass
+        else:
+            if hint_class is not sender:  # set in concrete
+                if not (self.inherit and issubclass(sender, hint_class)):  # set in abstract
+                    return
+        if hasattr(sender._meta, 'simple_history_manager_attribute'):
+            raise exceptions.MultipleRegistrationsError('{}.{} registered multiple times for history tracking.'.format(
+                sender._meta.app_label,
+                sender._meta.object_name,
+            ))
         history_model = self.create_history_model(sender)
         module = importlib.import_module(self.module)
         setattr(module, history_model.__name__, history_model)